Fix 32 bit off_t regression
The file_skip function was supposed to exactly test if off_t is 32 bit.
Eventually a refactoring removed this test.

Fixes: 7d2cbbc56c5db48705a23aa80d256fcd92d69e6f

+++ libarchive/archive_read_open_fd.c
@@ -131,7 +131,7 @@ static int64_t
 file_skip(struct archive *a, void *client_data, int64_t request)
 {
 	struct read_fd_data *mine = (struct read_fd_data *)client_data;
-	int64_t skip = request;
+	off_t skip = (off_t)request;
 	int64_t old_offset, new_offset;
 	int skip_bits = sizeof(skip) * 8 - 1;  /* off_t is a signed type. */
 
@@ -143,7 +143,7 @@ file_skip(struct archive *a, void *client_data, int64_t request)
 		int64_t max_skip =
 		    (((int64_t)1 << (skip_bits - 1)) - 1) * 2 + 1;
 		if (request > max_skip)
-			skip = max_skip;
+			skip = (off_t)max_skip;
 	}
 
 	/* Reduce request to the next smallest multiple of block_size */
